\name{estpab-package}
\alias{estpab-package}
\alias{estpab}
\docType{package}
\title{
\packageTitle{estpab}
}
\description{
\packageDescription{estpab}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{estpab}
\packageIndices{estpab}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{estpab}

Maintainer: \packageMaintainer{estpab}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
